+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.2.0]
+- feature: default height & width (plugin configuration)
+- feature: open edit mask after video selection
+- feature: edit mask improved (preview, slider, keep aspect ratio)
+
 ## [1.1.0]
 - ILIAS 5.4 support
 - bugfix: show own videos didn't work